feat(auth): Add authorization service to avoid edit/delete blogs an author doesnt own

+++ src/post/post.service.ts
@@ -3,10 +3,15 @@ import { PostStatus } from '@prisma/client';
 import { PrismaService } from '../prisma/prisma.service';
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
+import { RoleName } from '@prisma/client';
+import { AuthorizationService } from 'src/auth/services/authorization.service';
 
 @Injectable()
 export class PostService {
-  constructor(private readonly prisma: PrismaService) {}
+  constructor(
+    private readonly prisma: PrismaService,
+    private readonly authorizationService: AuthorizationService,
+  ) {}
 
   async create(dto: CreatePostDto, userId: string) {
     const slug = this.slugify(dto.title);
@@ -62,14 +67,21 @@ export class PostService {
     return post;
   }
 
-  async update(id: string, dto: UpdatePostDto) {
+  async update(
+    id: string,
+    dto: UpdatePostDto,
+    userId: string,
+    userRoles: RoleName[],
+  ) {
+    console.log('user id for update:', userId);
+    console.log('user roles for update:', userRoles);
+    await this.authorizationService.checkPostOwnership(id, userId, userRoles);
     await this.findOne(id);
 
     const { tags, ...rest } = dto;
     const data: any = { ...rest };
 
     if (tags) {
-      // Logic for handling tags update
       const tagCreateInput = this.processTags(tags);
       data.tagsOnPosts = {
         deleteMany: {},
@@ -92,7 +104,8 @@ export class PostService {
     });
   }
 
-  async remove(id: string) {
+  async remove(id: string, userId: string, userRoles: RoleName[]) {
+    await this.authorizationService.checkPostOwnership(id, userId, userRoles);
     await this.findOne(id);
     return this.prisma.post.delete({ where: { id } });
   }
